server/requirements.txt
Flask==3.1.0
flask-cors==5.0.1
python-dotenv==1.0.1
spotipy==2.25.1
lyricsgenius==3.2.0
numpy==2.0.2
pandas==2.2.3
requests==2.32.3
beautifulsoup4==4.12.2
accelerate==1.6.0
transformers 
openai==0.28.0
pillow==11.2.1 
torch==2.6.0 
torchvision==0.21.0 
tqdm==4.67.1 
transformers==4.51.3 
diffusers==0.33.1
replicate==1.0.4